+++ app.js
@@ -8,7 +8,7 @@ const morgan = require("morgan")
 const connect = require("./schemas")
 const passport = require("passport")
 const passportConfig = require("./routes/auth")
-// const { newsData } = require("./controllers/news.controller")
+const { newsData } = require("./services/news.service")
 
 connect()
 
@@ -66,7 +66,7 @@ app.use(passport.session())
 const Router = require("./routes")
 app.use("/api", Router)
 
-// newsData()
+newsData()
 
 app.use((req, res, next) => {
     res.status(404).send("요청하신 페이지를 찾을 수 없습니다.")
