+++ services/mark.service.js
@@ -7,6 +7,7 @@ redisSet = async (userId) => {
     const checkMark = await BokjiApi.find({ dataId: markInfo.mark }, { _id: false, dataId: true, name: true, desire: true })
     const redisInsertMark = JSON.stringify(checkMark)
     await Client.set(userId, redisInsertMark)
+    await Client.expire(userId, 7200)
 }
 
 exports.showMarkRedis = async (userId) => {
